requirements.txt
# Python dependencies for Qdrant upload script

qdrant-client>=1.7.0
fastembed>=0.2.0
python-dotenv>=1.0.0
